+++ hashmap.js
@@ -1,7 +1,7 @@
 /**
  * HashMap - HashMap Implementation for JavaScript
  * @author Jack Moxley <https://github.com/jackmoxley>
- * @version 0.5.0
+ * @version 0.6.0
  * Homepage: https://github.com/mootable/hashmap
  */
 
@@ -70,15 +70,16 @@
      *   forEach(func, ctx) : this
      */
     class HashMap {
-        constructor(iterable, _depth, _widthB) {
-            const widthB = _widthB ? _widthB : 6; // 2^6 = 64 buckets
-            const depth = _depth ?  _depth : ((32/widthB) >>> 0) - 1; // 0 indexed so 3 is a depth of 4.
+        constructor({copy, depth, widthB}= {}) {
+            widthB = Math.max(2, Math.min(16, widthB ? widthB : 6)); // 2^6 = 64 buckets
+            const defaultDepth = ((32/widthB) >> 0) - 1;
+            depth = Math.max(0, Math.min(depth && depth > 0 ? depth - 1 : defaultDepth, defaultDepth)); // 0 indexed so 3 is a depth of 4.
             const width = 1 << widthB; // 2 ^ widthB
             const mask = width - 1;
             this.options = Object.freeze({widthB, width, mask, depth});
             this.clear();
-            if (iterable) {
-                this.copy(iterable);
+            if (copy) {
+                this.copy(copy);
             }
         }
         has(key) {
@@ -130,7 +131,7 @@
             return this;
         }
         clone() {
-            return new HashMap(this, this.options);
+            return new HashMap({copy: this, depth: this.options.depth, widthB: this.options.widthB});
         }
         delete(key) {
             if(this.buckets) {
@@ -300,8 +301,8 @@
     HashMap.uid = 0;
 
     class LinkedHashMap extends HashMap {
-        constructor(iterable, _depth, _widthB) {
-            super(iterable, _depth, _widthB);
+        constructor(options) {
+            super(options);
             this.start = undefined;
             this.end = undefined;
         }
@@ -331,7 +332,7 @@
             return this;
         }
         clone() {
-            return new LinkedHashMap(this, this.options);
+            return new LinkedHashMap({copy: this, depth: this.options.depth, widthB: this.options.widthB});
         }
         forEach(func, ctx) {
             let entry = this.start;